LPS-124961 Move *ConfigurationBeanDeclaration to *.internal.settings.definition
